+++ packages/button/src/icon-button/types.ts
@@ -1,0 +1,16 @@
+export enum IconButtonSize {
+  XSmall = 'x-small',
+  Small = 'small',
+  Medium = 'medium',
+  Large = 'large',
+  XLarge = 'x-large',
+}
+
+export enum IconButtonColor {
+  Primary = 'primary',
+  Secondary = 'secondary',
+  Success = 'success',
+  Danger = 'danger',
+  Warning = 'warning',
+  Disabled = 'disabled',
+}