+++ packages/searchable-dropdown/src/dropdown/element.css.ts
@@ -2,53 +2,67 @@ import { css, unsafeCSS } from 'lit';
 import { styles as theme } from '@equinor/fusion-web-theme';
 import { styles as mdcStyle } from '@material/mwc-textfield/mwc-textfield.css';
 
-export const style = css`
+export const fwcsdd = css`
   :host {
+    position: relative;
     width: 100%;
+    --textinput-dense-size: 32px;
+    --fwc-text-field-base-color: ${unsafeCSS(theme.colors.text.static_icons__tertiary.getVariable('color'))};
+    --fwc-text-field-fill-color: ${unsafeCSS(theme.colors.ui.background__light.getVariable('color'))};
+    --fwc-text-field-ink-color: ${unsafeCSS(theme.colors.text.static_icons__default.getVariable('color'))};
+    --fwc-text-field-disabled-ink-color: ${unsafeCSS(theme.colors.text.static_icons__default.getVariable('color'))};
   }
-  fwc-searchable-dropdown {
-    width: 100%;
+  .input {
+    posistion: relative;
   }
   fwc-textinput {
     width: 100%;
   }
-  .fwc-sdd-column {
-    position: relative;
+  fwc-textinput[dense] {
+    --mdc-text-field-outlined-idle-border-color: transparent;
+    --mdc-shape-small: 0;
+    --mdc-text-field-outlined-hover-border-color: ${unsafeCSS(
+      theme.colors.interactive.primary__resting.getVariable('color')
+    )};
   }
-  .fwc-sdd-input {
-    display: flex;
-    align-items: center;
-    posistion: relative;
-  }
-  .no-hover {
-    pointer-events: none;
-  }
-  .trailing-slot {
-    position: absolute;
-    right: 5px;
-    top: 0;
-    height: calc(100% - 16px);
+  .interactive {
     cursor: pointer;
-    margin: 8px 0;
-    padding: 0 8px;
+  }
+  [slot='trailing'] {
+    display: block;
+    position: absolute;
+    top: 2px;
+    right: 2px;
+    bottom: 2px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: var(--textinput-dense-size);
     color: ${unsafeCSS(theme.colors.interactive.primary__resting.getVariable('color'))};
+    font-size: 0.8em;
   }
-  .trailing-slot:hover {
-    background: ${unsafeCSS(theme.colors.interactive.primary__hover_alt.getVariable('color'))};
-    border-radius: 50%;
-  }
-  .fwc-sdd-list {
+  .list {
     position: absolute;
     top: calc(100% + 4px);
-    right: 0;
     left: 0;
+    width: 100%;
     height: auto;
-    border-radius: 4px;
-    background-color: ${unsafeCSS(theme.colors.ui.background__light.getVariable('color'))};
-    box-shadow: 0px 1px 5px rgba(0, 0, 0, 0.2), 0px 3px 4px rgba(0, 0, 0, 0.12), 0px 2px 4px rgba(0, 0, 0, 0.14);
+    overflow: hidden;
     z-index: 99;
+    box-shadow: 0px 1px 5px rgba(0, 0, 0, 0.2), 0px 3px 4px rgba(0, 0, 0, 0.12), 0px 2px 4px rgba(0, 0, 0, 0.14);
+    border-radius: 4px;
   }
-  .fwc-sdd-outlined .fwc-sdd-list {
+  .list-scroll {
+    width: 100%;
+    height: auto;
+    overflow: hidden auto;
+    scrollbar-width: thin;
+    scrollbar-color: ${unsafeCSS(theme.colors.ui.background__medium.getVariable('color'))} transparent;
+  }
+  fwc-list {
+    background-color: ${unsafeCSS(theme.colors.ui.background__light.getVariable('color'))};
+  }
+  .variant-outlined fwc-list {
     background-color: ${unsafeCSS(theme.colors.ui.background__default.getVariable('color'))};
   }
   .item-text {
@@ -57,11 +71,16 @@ export const style = css`
     font-size: 1em;
   }
   .item-title {
-    height: 1.4em;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+    overflow: hidden;
     font-size: 1em;
     line-height: 1.6;
   }
   .item-subtitle {
+    text-overflow: ellipsis;
+    white-space: nowrap;
+    overflow: hidden;
     font-size: 0.8em;
     font-style: italic;
     line-height: 1.6;
@@ -87,6 +106,6 @@ export const style = css`
   }
 `;
 
-export const styles = [mdcStyle, style];
+export const styles = [mdcStyle, fwcsdd];
 
 export default styles;