+++ packages/searchable-dropdown/src/provider/controller.ts
@@ -1,14 +1,14 @@
 import { ReactiveController } from 'lit';
 import { Task } from '@lit-labs/task';
+
 import {
   SearchableDropdownResult,
   SearchableDropdownResolver,
   SearchableDropdownControllerHost,
   SearchableDropdownResultItem,
+  SearchableDropdownSelectEvent,
 } from '../types';
-import { SearchableDropdownConnectEvent } from '../events';
-import { ActionDetail } from '@material/mwc-list/mwc-list-foundation';
-import { v4 as uuid } from 'uuid';
+import { SearchableDropdownConnectEvent, ExplicitEventTarget } from '../types';
 
 export class SearchableDropdownController implements ReactiveController {
   protected disconnectProvider?: VoidFunction;
@@ -21,8 +21,10 @@ export class SearchableDropdownController implements ReactiveController {
   public result?: SearchableDropdownResult;
   public task!: Task<[string], SearchableDropdownResult>;
 
-  #queryString = '';
+  #externaCloseHandler?: (e: MouseEvent) => void;
   #host: SearchableDropdownControllerHost;
+  #queryString = '';
+  #initialItems: SearchableDropdownResult = [];
 
   constructor(host: SearchableDropdownControllerHost) {
     this.#host = host;
@@ -33,14 +35,22 @@ export class SearchableDropdownController implements ReactiveController {
     this.task = new Task<[string], SearchableDropdownResult>(
       this.#host,
       async ([qs]: [string]): Promise<SearchableDropdownResult> => {
-        if (!qs) {
-          return [{ id: uuid(), title: this.#host.initialText }];
-        }
         if (!this.resolver?.searchQuery) {
-          /* resolver is not setup the right way */
-          throw new Error('SeachableDropdownResolver is missing searchQuery handler');
+          /* resolver is not available */
+          return [];
         }
-        this.result = await this.resolver.searchQuery(qs);
+        let result;
+        if (!qs) {
+          if (this.#initialItems.length) {
+            result = this.#initialItems;
+          } else {
+            result = [{ id: 'initial', title: this.#host.initialText, isDisabled: true }];
+          }
+        } else {
+          result = await this.resolver.searchQuery(qs);
+        }
+        // set isSelected on result items
+        this.result = this.mutateResult(result);
         return this.result;
       },
       () => [this.#queryString]
@@ -49,10 +59,26 @@ export class SearchableDropdownController implements ReactiveController {
 
   protected updateResolver = (resolver?: SearchableDropdownResolver): void => {
     this.resolver = resolver;
+
+    if (resolver?.initialResult) {
+      this.#initialItems = resolver.initialResult;
+      this.task.run();
+    }
+
+    if (resolver?.closeHandler) {
+      this.#externaCloseHandler = resolver.closeHandler;
+    }
+
     this.#host.requestUpdate();
   };
 
   public hostConnected(): void {
+    requestAnimationFrame(() => {
+      if (this.#host.textInputElement && this.#host.autofocus) {
+        this.#host.textInputElement.focus();
+      }
+    });
+
     const event = new SearchableDropdownConnectEvent({
       detail: {
         disconnectedCallback: (callback) => {
@@ -65,6 +91,7 @@ export class SearchableDropdownController implements ReactiveController {
       cancelable: false,
     });
     this.#host.dispatchEvent(event);
+
     /* add click event to window */
     window.addEventListener('click', this._handleWindowClick);
   }
@@ -75,6 +102,7 @@ export class SearchableDropdownController implements ReactiveController {
     }
     /* remove click event to window */
     window.removeEventListener('click', this._handleWindowClick);
+    window.removeEventListener('keyup', this._handleWindowKeyUp);
   }
 
   /**
@@ -83,33 +111,91 @@ export class SearchableDropdownController implements ReactiveController {
   private _handleWindowClick = (e: Event): void => {
     /* make sure we have a target to check against */
     if (!e.target) return;
-    if ((e.target as HTMLElement).nodeName !== this.#host.nodeName) {
+    if (
+      (e.target as HTMLElement).nodeName !== this.#host.nodeName &&
+      document.activeElement?.nodeName !== this.#host.nodeName
+    ) {
       this.isOpen = false;
     }
   };
 
   /**
+   * Close dropdown on escape key
+   */
+  private _handleWindowKeyUp = (e: KeyboardEvent): void => {
+    /* Close on Escape */
+    if (e.key === 'Escape') {
+      if (this.#host.textInputElement) {
+        /* unfocus */
+        this.#host.textInputElement.blur();
+        /* Close element */
+        this.isOpen = false;
+      }
+    }
+  };
+
+  /**
+   * Mutates result to set parameters like isSelected.
+   * @param result SearchableDropdownResult
+   * @returns result
+   */
+  private mutateResult(result: SearchableDropdownResult) {
+    if (result) {
+      for (let i = 0; i < result.length; i++) {
+        const item = result[i];
+        if (item.type === 'section' && item.children?.length) {
+          for (let x = 0; x < item.children.length; x++) {
+            const kid = item.children[x];
+            if (this._selectedItems.find((s) => s.id === kid.id)) {
+              kid.isSelected = true;
+            } else {
+              kid.isSelected = false;
+            }
+          }
+        } else {
+          if (this._selectedItems.find((s) => s.id === item.id)) {
+            item.isSelected = true;
+          } else {
+            item.isSelected = false;
+          }
+        }
+      }
+    }
+
+    return result;
+  }
+
+  /**
    * Fires the select event to listener on host.
-   * using the action event from the fwc-list element.
-   * @param action Customevent with details property
+   * using the event event from the fwc-list element.
+   * @param event fwc-list action event with details property
    * @return SearchableDropdownResult the selected item in array.
    */
-  public handleSelect(action: CustomEvent<ActionDetail>): void {
-    action.stopPropagation();
+  // public handleSelect(event: CustomEvent<eventDetail>): void {
+  public handleSelect(event: ExplicitEventTarget): void {
+    event.stopPropagation();
+
+    /* dont fire select event when li child checkbox is clicked, for ex. a favourit checkbox */
+    if (event.explicitOriginalTarget && event.explicitOriginalTarget.type === 'checkbox') {
+      return;
+    }
 
     if (this.result && this._listItems) {
-      const id = this._listItems[action.detail.index];
+      const id = this._listItems[event.detail.index];
 
       /* Find selected item in resolver result list */
       let selectedItem: SearchableDropdownResultItem | undefined;
+
       // get selected item from result
       for (const item of this.result) {
         if (item.id === id) {
           selectedItem = item;
+          break;
         } else if (item.children) {
           for (const childItem of item.children) {
             if (childItem.id === id) {
               selectedItem = childItem;
+              break;
             }
           }
         }
@@ -117,32 +203,41 @@ export class SearchableDropdownController implements ReactiveController {
 
       /* Set Error if none matched the resolver result */
       if (!selectedItem?.id) {
-        throw new Error(
-          'SearchableDropdownControlloer could not find match  in selectedItem to result provided by resolver.'
-        );
+        throw new Error('SearchableDropdownController could not find match  in result provided by resolver.');
       }
 
       /*  Set active state and save selected item in state */
-      if (this._selectedItems.find((si) => si.id === selectedItem?.id)) {
-        /*  Already selected so clear it from selections */
-        selectedItem.isSelected = false;
-        this._selectedItems = this._selectedItems.filter((i) => i.id !== selectedItem?.id);
-        this.#host.selected = '';
+      if (this.#host.multiple) {
+        if (this._selectedItems.find((si) => si.id === selectedItem?.id)) {
+          /*  Already selected so clear it from selections */
+          selectedItem.isSelected = false;
+          this._selectedItems = this._selectedItems.filter((i) => i.id !== selectedItem?.id);
+          this.#host.value = '';
+        } else {
+          /*  Adds new item to selections */
+          selectedItem.isSelected = true;
+          this._selectedItems.push(selectedItem);
+          this.#host.value = selectedItem?.title || '';
+        }
       } else {
         /*  Adds new item to selections */
-        selectedItem.isSelected = true;
-        this._selectedItems.push(selectedItem);
-        this.#host.selected = selectedItem?.title || '';
+        this._selectedItems = [selectedItem];
+        this.#host.value = selectedItem?.title || '';
       }
     } else {
+      /* FALSE === this.result && this._listItems */
       /* Clear selected states */
       this._selectedItems = [];
-      this.#host.selected = '';
+      this.#host.value = '';
+    }
+
+    if (!this.#host.multiple) {
+      this.isOpen = false;
     }
 
     /* Dispatch custom select event with our details */
     this.#host.dispatchEvent(
-      new CustomEvent('select', {
+      new SearchableDropdownSelectEvent({
         detail: {
           selected: this._selectedItems,
         },
@@ -150,15 +245,65 @@ export class SearchableDropdownController implements ReactiveController {
       })
     );
 
+    /* Sets items isSelected in task */
+    this.task.run();
+
     /* Refresh host */
     this.#host.requestUpdate();
   }
 
+  /**
+   * Fires on click on close icon in textinput
+   * Calls closeHandler callback set in resolver
+   */
+  public closeClick = (e: MouseEvent): void => {
+    /* needed to clear user input */
+    if (this.#host.textInputElement) {
+      this.#host.textInputElement.value = '';
+      this.#host.textInputElement.blur();
+    }
+
+    this.#host.value = '';
+    this._selectedItems = [];
+    this.#queryString = '';
+
+    /* also runs task */
+    this.isOpen = false;
+
+    /* call resolvers handleclick if defined */
+    if (this.#externaCloseHandler) {
+      this.#externaCloseHandler(e);
+    }
+
+    /* fire event for sdd closed */
+    const ddClosedEvent = new CustomEvent<{ date: number }>('dropdownClosed', {
+      detail: {
+        date: Date.now(),
+      },
+      bubbles: true,
+    });
+    this.#host.dispatchEvent(ddClosedEvent);
+  };
+
   /* Settter: Open/Closed state for host */
   public set isOpen(state: boolean) {
-    this.#queryString = '';
     this._isOpen = state;
+    // toogle close icon
     this.#host.trailingIcon = state ? 'close' : '';
+
+    /* syncs dropdown list with textinput */
+    if (this._selectedItems) {
+      this.task.run();
+    }
+
+    /* Close on escape key */
+    if (this._isOpen) {
+      window.addEventListener('keyup', this._handleWindowKeyUp);
+    } else {
+      window.removeEventListener('keyup', this._handleWindowKeyUp);
+    }
+
+    /* Refresh host */
     this.#host.requestUpdate();
   }
 
@@ -175,7 +320,8 @@ export class SearchableDropdownController implements ReactiveController {
     const target = event.target as HTMLInputElement;
     if (event.key === 'ArrowDown') {
       if (this._isOpen && this.result?.length) {
-        this.#host.renderRoot.querySelector('fwc-list').focus();
+        // focus on the fwc-list
+        this.#host.renderRoot.querySelector('fwc-list')?.focus();
       }
       return;
     }
@@ -184,7 +330,6 @@ export class SearchableDropdownController implements ReactiveController {
     }
     this.timer = setTimeout(() => {
       this.#queryString = target.value.trim().toLowerCase();
-      // this.host.pendingQuery = this._searchQueryTask(this.resolver);
       this.#host.requestUpdate();
     }, 250);
   }