+++ storybook/stories/Components/SearchableDropdown.tsx
@@ -11,30 +11,31 @@ import {
 SearchableDropdownElement;
 SearchableDropdownProviderElement;
 
-/* 
-import json from '../resources/context.json';
-// map items to SearchableDropdownResult
-const allItems: SearchableDropdownResult = json.map((item) => {
-  return {
-    id: item.id,
-    title: item.title,
-    subTitle: item.type.id,
-    // isError: item.isDeleted,
-    // isSelected: !item.isActive,
-    // type: item.type,
-    // children: item.children,
-  };
-});
-*/
-import allItems from '../resources/sections.json';
+// import allItems from '../resources/sections.json';
+
+import rawItems from '../resources/context.json';
+
+const mapper = (src: Array<{ id: string; title: string; type: { id: string } }>): SearchableDropdownResult => {
+  const dst = src.map((i) => {
+    return {
+      id: i.id,
+      title: i.title,
+      // subTitle: i.type.id,
+      graphic: i.type.id === 'OrgChart' ? 'list' : undefined,
+    };
+  });
+
+  return dst;
+};
+
+const allItems = mapper(rawItems);
 
 /* generate single SearchableDropdownResult item */
-const singleItem = (props: unknown): SearchableDropdownResultItem => {
+const singleItem = (props: Partial<SearchableDropdownResultItem>): SearchableDropdownResultItem => {
   return Object.assign({ id: '0', title: 'Dummy title' }, props);
 };
 
 /* Dummy api handler */
-
 const apiItems = (query: string): SearchableDropdownResult => {
   /* min length of query string */
   const min = 2;
@@ -47,6 +48,7 @@ const apiItems = (query: string): SearchableDropdownResult => {
   /* Recursive func for matching in children  */
   for (const item of allItems as SearchableDropdownResult) {
     const entry = { ...item };
+    // Match against children in sections
     if (entry.type === 'section' && entry.children?.length) {
       entry.children = entry.children.filter((i) => i.title && i.title.toLowerCase().indexOf(query) > -1);
       if (entry.children.length) {
@@ -65,13 +67,40 @@ const apiItems = (query: string): SearchableDropdownResult => {
 };
 
 const resolver: SearchableDropdownResolver = {
-  searchQuery: async (query: string) => {
+  searchQuery: (query: string) => {
     try {
       // Dummy api call returning matches
       return apiItems(query);
     } catch {
       return [singleItem({ title: 'Error while searcing', isDisabled: true, isError: true })];
     }
+  },
+  initialResult: [
+    singleItem({
+      id: 'ctx-123',
+      title: 'Context',
+      type: 'section',
+      children: [
+        singleItem({ id: '456', title: 'Context 1', graphic: 'list' }),
+        singleItem({ id: '654', title: 'Context 2', graphic: 'list' }),
+        singleItem({ id: '789', title: 'Context 3', graphic: 'list' }),
+        singleItem({ id: '321', title: 'Context 4', graphic: 'list' }),
+      ],
+    }),
+    singleItem({
+      id: 'fav-123',
+      title: 'Favourites',
+      type: 'section',
+      children: [
+        singleItem({ id: '456456', title: 'Favourite 1', meta: 'check' }),
+        singleItem({ id: '654654', title: 'Favourite 2', meta: 'check' }),
+        singleItem({ id: '789789', title: 'Favourite 3', meta: 'check' }),
+        singleItem({ id: '321321', title: 'Favourite 4', meta: 'check' }),
+      ],
+    }),
+  ],
+  closeHandler: (e: MouseEvent) => {
+    console.log('closeEvent fired', e);
   },
 };
 
@@ -81,11 +110,12 @@ const useSearchableDropdownProviderRef = (
   const providerRef = useRef<SearchableDropdownProviderElement>(null);
   useEffect(() => {
     if (providerRef?.current) {
+      providerRef.current.connectResolver(resolver);
       providerRef?.current.addEventListener('select', (e) => console.log('Event', e));
-      providerRef.current.setResolver(resolver);
+      providerRef?.current.addEventListener('dropdownClosed', (e) => console.log('Event', e));
       return () => {
-        providerRef.current?.removeEventListener('select', (e) => console.log('Event', e));
         providerRef.current?.removeResolver();
+        providerRef.current?.removeEventListener('select', (e) => console.log('Event', e));
       };
     }
   }, [providerRef]);
@@ -98,7 +128,7 @@ export const SearchableDropdown = ({ children, ...props }: PropsWithChildren<Sea
 
   return (
     <fwc-searchable-dropdown-provider ref={providerRef}>
-      <fwc-searchable-dropdown {...extractProps<SearchableDropdownProps>(props)}>{children}</fwc-searchable-dropdown>
+      <fwc-searchable-dropdown {...extractProps(props)}>{children}</fwc-searchable-dropdown>
     </fwc-searchable-dropdown-provider>
   );
 };