+++ storybook/stories/Components/Skeleton.tsx
@@ -1,4 +1,4 @@
-import { PropsWithChildren } from 'react';
+import React, { PropsWithChildren } from 'react';
 
 import extractProps from './extract-props';