+++ packages/divider/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.18](https://github.com/equinor/fusion-web-components/compare/@equinor/fusion-wc-divider@0.2.17...@equinor/fusion-wc-divider@0.2.18) (2022-11-24)
+
+**Note:** Version bump only for package @equinor/fusion-wc-divider
+
+
+
+
+
 ## [0.2.17](https://github.com/equinor/fusion-web-components/compare/@equinor/fusion-wc-divider@0.2.16...@equinor/fusion-wc-divider@0.2.17) (2022-11-07)
 
 **Note:** Version bump only for package @equinor/fusion-wc-divider