+++ packages/list/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.1](https://github.com/equinor/fusion-web-components/compare/@equinor/fusion-wc-list@0.3.0...@equinor/fusion-wc-list@0.3.1) (2022-11-24)
+
+**Note:** Version bump only for package @equinor/fusion-wc-list
+
+
+
+
+
 # [0.3.0](https://github.com/equinor/fusion-web-components/compare/@equinor/fusion-wc-list@0.2.2...@equinor/fusion-wc-list@0.3.0) (2022-11-08)